batchMapQuick {BatchJobs}	R Documentation

Combination of makeRegistry, batchMap and submitJobs.

Description

Combination of makeRegistry, batchMap and submitJobs for quick computations on the cluster. Should only be used by skilled users who know what they are doing. Creates the file.dir in current directory under the name “bmq_[random alphanumerics]”, maps function, potentially chunks jobs and submits them.

Usage

  batchMapQuick(fun, ..., more.args = list(),
    packages = character(0L), chunk.size, n.chunks,
    chunks.as.arrayjobs = FALSE, inds, resources = list(),
    temporary = FALSE)

Arguments

fun
[function]
Function to map over ....

...
[any]
Arguments to vectorize over (list or vector).

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

packages
[character]
Packages that will always be loaded on each node. Default is character(0).

chunk.size
[integer(1)]
Preferred number of jobs in each chunk. Can not be used in combination with n.chunks. Note that the ids will get shuffled to balance out possible run time differences. Default is not to use chunking.

n.chunks
[integer(1)]
Preferred number chunks. Can not be used in combination with chunk.size. Note that the ids will get shuffled to balance out possible run time differences. Default is not to use chunking.

chunks.as.arrayjobs
[logical(1)]
Submit chunks as array jobs? Default is FALSE.

inds
[integer]
Indices of ids / chunks to submit. Default is all.

resources
[list]
Required resources for all batch jobs. Default is empty list.

temporary
[logical(1)]
Create the file.dir inside R's tempdir? Default is FALSE.

Value

[Registry]

[Package BatchJobs version 1.2 Index]