batchMapResults {BatchJobs}	R Documentation

Maps a function over the results of a registry by using batchMap.

Description

Maps a function over the results of a registry by using batchMap.

Usage

batchMapResults(reg, reg2, fun, ..., ids, part = NA_character_,
  more.args = list())

Arguments

reg
[Registry]
Registry whose results should be mapped by fun.

reg2
[Registry]
Empty registry that should store the job for the mapping.

fun
[function(job, res, ...)]
Function to map over results of reg. Further arguments come from ... of batchMapResults and more.args.

...
[any]
Furher arguments to vectorize over (list or vector). Must all be the same length as number of results in reg.

ids
[integer]
Ids of jobs whose results should be mapped with fun. Default is all jobs.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

Value

Vector of type integer with job ids.

Examples

reg1 = makeRegistry(id = "BatchJobsExample1", file.dir = tempfile(), seed = 123)

## Creating dir: /tmp/RtmpxaPSRM/file709f571f0e7
## Saving registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData

# square some numbers
f = function(x) x^2
batchMap(reg1, f, 1:10)

## Adding 10 jobs to DB.

# submit jobs and wait for the jobs to finish
submitJobs(reg1)

## Saving conf: /tmp/RtmpxaPSRM/file709f571f0e7/conf.RData
## Submitting 10 chunks / 10 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 10 R scripts...
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:34: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-07-10 07:49:34
## Setting seed: 123
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/01/1-result.RData
## 2014-07-10 07:49:34: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 124
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/02/2-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 125
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/03/3-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 126
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/04/4-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 127
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/05/5-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 128
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/06/6-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 129
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/07/7-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 130
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/08/8-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 131
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/09/9-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f571f0e7/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 132
## Writing result file: /tmp/RtmpxaPSRM/file709f571f0e7/jobs/10/10-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Sending 10 submit messages...
## Might take some time, do not interrupt this!

# look at results
reduceResults(reg1, fun = function(aggr,job,res) c(aggr, res))

## Reducing

##  [1]   1   4   9  16  25  36  49  64  81 100

reg2 = makeRegistry(id = "BatchJobsExample2", file.dir = tempfile(), seed = 123)

## Creating dir: /tmp/RtmpxaPSRM/file709f5a2169d
## Saving registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData

# define function to tranform results, we simply do the inverse of the squaring
g = function(job, res) sqrt(res)
batchMapResults(reg1, reg2, fun = g)

## Saving registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Adding 10 jobs to DB.

# submit jobs and wait for the jobs to finish
submitJobs(reg2)

## Saving conf: /tmp/RtmpxaPSRM/file709f5a2169d/conf.RData
## Submitting 10 chunks / 10 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 10 R scripts...
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 123
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/01/1-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 124
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/02/2-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 125
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/03/3-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:35: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-07-10 07:49:35
## Setting seed: 126
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/04/4-result.RData
## 2014-07-10 07:49:35: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:36: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-07-10 07:49:36
## Setting seed: 127
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/05/5-result.RData
## 2014-07-10 07:49:36: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:36: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-07-10 07:49:36
## Setting seed: 128
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/06/6-result.RData
## 2014-07-10 07:49:36: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:36: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-07-10 07:49:36
## Setting seed: 129
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/07/7-result.RData
## 2014-07-10 07:49:36: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:36: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-07-10 07:49:36
## Setting seed: 130
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/08/8-result.RData
## 2014-07-10 07:49:36: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:36: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-07-10 07:49:36
## Setting seed: 131
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/09/9-result.RData
## 2014-07-10 07:49:36: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpxaPSRM/file709f5a2169d/registry.RData
## Loading conf: 
## 2014-07-10 07:49:36: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-07-10 07:49:36
## Setting seed: 132
## Writing result file: /tmp/RtmpxaPSRM/file709f5a2169d/jobs/10/10-result.RData
## 2014-07-10 07:49:36: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Sending 10 submit messages...
## Might take some time, do not interrupt this!

# check results
reduceResults(reg2, fun = function(aggr,job,res) c(aggr, res))

## Reducing

##  [1]  1  2  3  4  5  6  7  8  9 10

[Package BatchJobs version 1.3 Index]