cfBrewTemplate {BatchJobs}	R Documentation

Cluster functions helper: Brew your template into a job description file.

Description

This function is only intended for use in your own cluster functions implementation.

Usage

cfBrewTemplate(conf, template, rscript, extension)

Arguments

conf
[environment]
BatchJobs configuration.

template
[character(1)]
Job desfription template as a char vecrtor, possibly read in via cfReadBrewTemplate.

rscript
[character(1)]
File path to you your corresponding R script for the job.

extension
[character(1)]
Extension for the job description file, e.g. “pbs”.

Details

Calls brew silently on your template, any error will lead to an exception. If debug mode is turned on in the configuration, the file is stored at the same place as the corresponding R script in the “jobs”-subdir of your files directory, otherwise in the temp dir via tempfile.

Value

[character(1)]. File path of result.

[Package BatchJobs version 1.3 Index]