cfHandleUnknownSubmitError {BatchJobs}	R Documentation

Cluster functions helper: Handle an unknown error during job submission.

Description

This function is only intended for use in your own cluster functions implementation.

Usage

cfHandleUnknownSubmitError(cmd, exit.code, output)

Arguments

cmd
[character(1)]
OS command used to submit the job, e.g. qsub.

exit.code
[integer(1)]
Exit code of the OS command, should not be 0.

output
[character]
Output of the OS command, hopefully an informative error message. If these are mutiple lines in a vector, they are automatically pasted together.

Details

Simply constructs a SubmitJobResult object with status code 101, NA as batch job id and an informative error message containing the output of the OS command in output.

Value

[SubmitJobResult].

[Package BatchJobs version 1.3 Index]