cfKillBatchJob {BatchJobs}	R Documentation

Cluster functions helper: Kill a batch job via OS command

Description

This function is only intended for use in your own cluster functions implementation.

Usage

cfKillBatchJob(cmd, batch.job.id, max.tries = 3L)

Arguments

cmd
[character(1)]
OS command, e.g. “qdel”.

batch.job.id
[character(1)]
Id of the batch job on the batch system.

max.tries
[integer(1)]
Number of total times to try execute the OS command in cases of failures. Default is 3.

Details

Calls the OS command to kill a job via system like this: “cmd batch.job.id”. If the command returns an exit code > 0, the command is repeated after a 1 second sleep max.tries-1 times. If the command failed in all tries, an exception is generated.

Value

Nothing.

[Package BatchJobs version 1.3 Index]