cfReadBrewTemplate {BatchJobs}	R Documentation

Cluster functions helper: Read in your brew template file.

Description

This function is only intended for use in your own cluster functions implementation.

Usage

cfReadBrewTemplate(template.file)

Arguments

template.file
[character(1)]
File path.

Details

Simply reads your template and returns it as a character vector. If you do this in the constructor of your cluster functions once, you can avoid this repeated file access later on.

Value

[character].

[Package BatchJobs version 1.3 Index]