debugMulticore {BatchJobs}	R Documentation

Helper function to debug multicore mode.

Description

Useful in case of severe errors. Tries different operations of increasing difficulty and provides debug output on the console

Usage

debugMulticore(r.options)

Arguments

r.options
[list] Options for R and Rscript, one option per element of the vector, a la “–vanilla”. Default is c("--no-save", "--no-restore", "--no-init-file", "--no-site-file").

Value

Nothing.

[Package BatchJobs version 1.3 Index]