debugSSH {BatchJobs}	R Documentation

Helper function to debug SSH mode.

Description

Useful in case of configuration problems. Tries different operations of increasing difficulty and provides debug output on the console.

Usage

  debugSSH(nodename, rhome = "",
    r.options = c("--no-save", "--no-restore", "--no-init-file", "--no-site-file"),
    dir = getwd())

Arguments

nodename
[character(1)]
Node on which worker should be constructed for the test.

rhome
[character(1)]
Path to R installation on the worker. “” means R installation on the PATH is used, of course this implies that it must be on the PATH (also for non-interactive shells)! Default is “”.

r.options
[list] Options for R and Rscript, one option per element of the vector, a la “–vanilla”. Default is c("--no-save", "--no-restore", "--no-init-file", "--no-site-file").

dir
[character(1)]
Path where internally used test registries can be created. Note that this must be shared for the worker. Default is current working directory.

Value

Nothing.

[Package BatchJobs version 1.2 Index]