filterResults {BatchJobs}	R Documentation

Find all results where a specific condition is true.

Description

Find all results where a specific condition is true.

Usage

filterResults(reg, ids, fun, ...)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs whose results you want to test for the condition. Default is all jobs for which results are available.

fun
[fun(job, res)]
Predicate function that returns TRUE or FALSE.

...
[any]
Additional arguments to fun.

Value

[integer]. Ids of jobs where fun(job, result) returns TRUE.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)
f <- function(x) x^2
batchMap(reg, f, 1:10)
submitJobs(reg)

# which square numbers are even:
filterResults(reg, fun=function(job, res) res %% 2 == 0)

## [1]  2  4  6  8 10

[Package BatchJobs version 1.3 Index]