findJobs {BatchJobs}	R Documentation

Finds ids of jobs that match a query.

Description

Finds ids of jobs that match a query.

Usage

findJobs(reg, ids, pars, jobnames)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Subset of job ids to restrict the result to. Default is all jobs.

pars
[R expression]
All jobs whose parameters match the given expression are selected. This implies that you have named the parameters when you passed the vectors. If you forgot to do this you can use .arg1, .arg2, etc., to refer to the the unnamed ones.

jobnames
[character]
Restrict to jobs with stored names. Exact matching is used.

Value

[integer]. Ids for jobs which match the query.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)
f <- function(x, y) x * y
batchExpandGrid(reg, f, x=1:2, y=1:3)

##   x y
## 1 1 1
## 2 2 1
## 3 1 2
## 4 2 2
## 5 1 3
## 6 2 3

findJobs(reg, pars=(y > 2))

## [1] 5 6

[Package BatchJobs version 1.3 Index]