findDone {BatchJobs}	R Documentation

Find jobs depending on computional state.

Description

findDone: Find jobs which succesfully terminated.

findNotDone: Find jobs for which results are still missing.

findMissingResults: Deprecated. Alias for findNotDone.

findErrors: Find jobs where errors occured.

findNotErrors: Find jobs where no errors occured.

findTerminated: Find jobs which have terminated (done / error).

findNotTerminated: Find jobs which have not terminated (not done / no error).

findSubmitted: Find jobs which have been submitted.

findNotSubmitted: Find jobs which have not been submitted.

findOnSystem: Find jobs which are present on the batch system at the moment.

findNotOnSystem: Find jobs which are not present on the batch system at the moment.

findRunning: Find jobs which are running.

findNotRunning: Find jobs which are not running.

findStarted: Find jobs which have been started on the batch system.

findStarted: Find jobs which have not been started on the batch system. findNotRunning: Find jobs which are not running.

findExpired: Find jobs where walltime was probably hit. Right now the heuristic is as follows: Find all jobs that have submitted, did not abort with an error, did not complete with a result and are not submitted or running anymore.

Usage

findDone(reg, ids)

findNotDone(reg, ids)

findMissingResults(reg, ids)

findErrors(reg, ids)

findNotErrors(reg, ids)

findTerminated(reg, ids)

findNotTerminated(reg, ids)

findSubmitted(reg, ids)

findNotSubmitted(reg, ids)

findOnSystem(reg, ids)

findNotOnSystem(reg, ids)

findRunning(reg, ids)

findNotRunning(reg, ids)

findStarted(reg, ids)

findNotStarted(reg, ids)

findExpired(reg, ids)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Subset of job ids to restrict the result to. Default is all jobs.

Value

[integer]. Ids of jobs.

[Package BatchJobs version 1.3 Index]