getJobInfo {BatchJobs}	R Documentation

Get computational information of jobs.

Description

Returns time stamps (submitted, started, done), time running, time in queue, error messages (shortened, see showLog for detailed error messages), hostname of the host the job was executed, the assigned batch ID, the R PID and the seed of the job.

Usage

  getJobInfo(reg, ids, pars = FALSE, prefix.pars = FALSE,
    select, unit = "seconds")

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs. Default is all jobs.

pars
[logical(1)]
Include job parameters in the output? Default is FALSE.

prefix.pars
[logical(1)]
Should a prefix be added to job parameter names (column names) to avoid name clashes? Default is FALSE.

select
[character]
Select only a subset of columns. Usually this is not required and you can subset yourself, but in some rare cases it may be advantageous to not query all information. Note that the column “id” (job id) is always selected. If not provided, all columns are queried and returned.

unit
[character(1)]
Unit to convert execution and queing times to. Possible values: “seconds”, “minutes”, “hours”, “days” and “weeks”. Default is “seconds”.

Value

[data.frame].

[Package BatchJobs version 1.2 Index]