getJobResources {BatchJobs}	R Documentation

Function to get the resources that were submitted for some jobs.

Description

Throws an error if call it for unsubmitted jobs.

Usage

  getJobResources(reg, ids, as.list = TRUE)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs. Default is all submitted jobs.

as.list
[integer(1)]
If FALSE, rbind.fill is called on the result to convert it to a data.frame. Default is TRUE.

Value

[list | data.frame]. List (or data.frame) of resource lists as passed to submitJobs.

[Package BatchJobs version 1.2 Index]