getSSHWorkersInfo {BatchJobs}	R Documentation

Print and return R installation and other information for SSH workers.

Description

Workers are queried in parallel via callFunctionOnSSHWorkers.

Usage

  getSSHWorkersInfo(nodenames)

Arguments

nodenames
[character]
Nodenames of workers.

Details

The function will display a warning if the first lib path on the worker is not writable as this indicates potential problems in the configuration and installPackagesOnSSHWorkers will not work.

Value

[list]. Displayed information as a list named by nodenames.

See Also

callFunctionOnSSHWorkers

[Package BatchJobs version 1.2 Index]