grepLogs {BatchJobs}	R Documentation

Grep log files for a pattern.

Description

Searches for occurence of pattern in log files.

Usage

  grepLogs(reg, ids, pattern = "warn", ignore.case = TRUE,
    verbose = FALSE, range = 2L)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs to grep. Default is all terminated jobs (done + errors).

pattern
[character(1)]
Pattern to search for. See grep. Default is "warn".

ignore.case
[logical(1)]
Ignore case. See grep. Default is TRUE.

verbose
[logical(1)]
Print matches. Default is FALSE.

range
[integer(1)]
If verbose is set to TRUE, print range leading and trailing lines for contextual information about the warning. Default is 2.

Value

[integer]. Ids of jobs where pattern was found in the log file.

See Also

showLog, getErrorMessages

[Package BatchJobs version 1.2 Index]