installPackagesOnSSHWorkers {BatchJobs}	R Documentation

Install packages on SSH workers.

Description

Installation is done via callFunctionOnSSHWorkers and install.packages.

Usage

installPackagesOnSSHWorkers(nodenames, pkgs, repos = getOption("repos"),
  consecutive = TRUE, show.output = consecutive, ...)

Arguments

nodenames
[character]
Nodenames of workers.

pkgs
[character]
See install.packages.

repos
[character]
See install.packages. If the user must be queried this is of course done on the master.

consecutive
[logical(1)]
See callFunctionOnSSHWorkers. Default is TRUE.

show.output
[logical(1)]
See callFunctionOnSSHWorkers. Default is consecutive.

...
[any]
Passed to install.packages.

Details

Note that as usual the function tries to install the packages into the first path of .libPaths() of each each worker.

Value

Nothing.

See Also

callFunctionOnSSHWorkers

[Package BatchJobs version 1.3 Index]