killJobs {BatchJobs}	R Documentation

Kill some jobs on the batch system.

Description

Kill jobs which have already been submitted to the batch system. If a job is killed its internal state is reset as if it had not been submitted at all.

Usage

killJobs(reg, ids)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs to kill. Default is none.

Details

The function informs if (a) the job you want to kill has not been submitted, (b) the job has already terminated, (c) for some reason no batch job id is available. In all 3 cases above, nothing is changed for the state of this job and no call to the internal kill cluster function is generated.

In case of an error when killing, the function tries - after a short sleep - to kill the remaining batch jobs again. If this fails again for some jobs, the function gives up. Only jobs that could be killed are reset in the DB.

Value

[integer]. Ids of killed jobs.

Examples

## Not run: 
##D reg = makeRegistry(id = "BatchJobsExample", file.dir = tempfile(), seed = 123)
##D f = function(x) Sys.sleep(x)
##D batchMap(reg, f, 1:10 + 5)
##D submitJobs(reg)
##D 
##D # kill all jobs currently _running_
##D killJobs(reg, findRunning(reg))
##D # kill all jobs queued or running
##D killJobs(reg, findNotTerminated(reg))
## End(Not run)

[Package BatchJobs version 1.3 Index]