loadExports {BatchJobs}	R Documentation

Load exported R data objects stored.

Description

Loads all RData files in the “exports” subdirectory of your file.dir and assigns the objects to the global environment.

Usage

loadExports(reg)

Arguments

reg
[Registry]
Location of the file.dir to load the registry from.

Value

[character]. Invisibly returns a character vector of loaded file names.

[Package BatchJobs version 1.3 Index]