loadResult {BatchJobs}	R Documentation

Loads a specific result file.

Description

Loads a specific result file.

Usage

  loadResult(reg, id, part = NA_character_,
    missing.ok = FALSE)

Arguments

reg
[Registry]
Registry.

id
[integer(1)]
Id of job.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

missing.ok
[logical(1)]
If FALSE an error is thrown if no result file is found. Otherwise NULL is returned. Default is FALSE.

Value

[any]. Result of job.

See Also

reduceResults

[Package BatchJobs version 1.2 Index]