loadResults {BatchJobs}	R Documentation

Loads result files for id vector.

Description

Loads result files for id vector.

Usage

  loadResults(reg, ids, part = NA_character_,
    simplify = FALSE, use.names = "ids",
    missing.ok = FALSE)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs. Default is all done jobs.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

simplify
[logical(1)]
Should the result be simplified to a vector, matrix or higher dimensional array if possible? Default is TRUE.

use.names
[character(1)]
Name the results with job ids (“ids”), stored job names (“names”) or return a unnamed result (“none”). Default is ids.

missing.ok
[logical(1)]
If FALSE an error is thrown if the results are not found. Otherwise missing results are imputed to NULL. Default is FALSE.

Value

[list]. Results of jobs as list, possibly named by ids.

See Also

reduceResults

[Package BatchJobs version 1.2 Index]