makeClusterFunctionsInteractive {BatchJobs}	R Documentation

Create cluster functions for sequential execution in same session.

Description

All jobs executed under these cluster functions are executed sequentially, in the same interactive R process that you currently are. That is, submitJob does not return until the job has finished. The main use of this ClusterFunctions implementation is to test and debug programs on a local computer.

Usage

  makeClusterFunctionsInteractive(write.logs = TRUE)

Arguments

write.logs
[logical(1)]
Sink the output to log files. Turning logging off can increase the speed of calculations but makes it next to impossible to debug. Default is TRUE.

Details

Listing jobs returns an empty vector (as no jobs can be running when you call this) and killJob returns at once (for the same reason).

Value

[ClusterFunctions].

[Package BatchJobs version 1.2 Index]