makeClusterFunctionsLSF {BatchJobs}	R Documentation

Create cluster functions for LSF systems.

Description

Job files are created based on the brew template template.file. This file is processed with brew and then submitted to the queue using the bsub command. Jobs are killed using the bkill command and the list of running jobs is retrieved using bjobs -u $USER -w. The user must have the appropriate privileges to submit, delete and list jobs on the cluster (this is usually the case).

Usage

makeClusterFunctionsLSF(template.file)

Arguments

template.file
[character(1)]
Path to a brew template file that is used for the job file.

Details

The template file can access all arguments passed to the submitJob function, see here ClusterFunctions. It is the template file's job to choose a queue for the job and handle the desired resource allocations. Examples can be found on https://github.com/tudo-r/BatchJobs/tree/master/examples/cfLSF.

Value

[ClusterFunctions].

[Package BatchJobs version 1.3 Index]