makeClusterFunctionsLocal {BatchJobs}	R Documentation

Create cluster functions for sequential execution on local host.

Description

All jobs executed under these cluster functions are executed sequentially, but in an independent, new R session. That is, submitJob does not return until the job has finished. The main use of this ClusterFunctions implementation is to test and debug programs on a local computer.

Usage

  makeClusterFunctionsLocal()

Details

Listing jobs returns an empty vector (as no jobs can be running when you call this) and killJob returns at once (for the same reason).

Value

[ClusterFunctions].

[Package BatchJobs version 1.2 Index]