makeClusterFunctionsSGE {BatchJobs}	R Documentation

Create cluster functions for Sun Grid Engine systems.

Description

Job files are created based on the brew template template.file. This file is processed with brew and then submitted to the queue using the qsub command. Jobs are killed using the qdel command and the list of running jobs is retrieved using qselect. The user must have the appropriate privileges to submit, delete and list jobs on the cluster (this is usually the case).

The template file can access all arguments passed to the submitJob function, see here ClusterFunctions. It is the template file's job to choose a queue for the job and handle the desired resource allocations. Examples can be found on https://github.com/tudo-r/BatchJobs/tree/master/examples/cfSGE.

Usage

makeClusterFunctionsSGE(template.file, list.jobs.cmd = c("qstat", "-u $USER"))

Arguments

template.file
[character(1)]
Path to a brew template file that is used for the PBS job file.

list.jobs.cmd
[character]
Change default system command / options to list jobs. The first entry is the command, the following the options. See system3.

Value

[ClusterFunctions].

[Package BatchJobs version 1.3 Index]