makeClusterFunctionsSLURM {BatchJobs}	R Documentation

Create cluster functions for SLURM-based systems.

Description

Job files are created based on the brew template template.file. This file is processed with brew and then submitted to the queue using the sbatch command. Jobs are killed using the scancel command and the list of running jobs is retrieved using squeue. The user must have the appropriate privileges to submit, delete and list jobs on the cluster (this is usually the case).

Usage

  makeClusterFunctionsSLURM(template.file)

Arguments

template.file
[character(1)]
Path to a brew template file that is used for the SLURM job file.

Details

The template file can access all arguments passed to the submitJob function, see here ClusterFunctions. It is the template file's job to choose a queue for the job and handle the desired resource allocations. A simple example is provided here http://code.google.com/p/batchjobs/source/browse/trunk/BatchJobs/examples/cfSLURM/simple.tmpl and a more complex one here http://code.google.com/p/batchjobs/source/browse/trunk/BatchJobs/examples/cfSLURM/dortmund_fk_statistik.tmpl in the package repository on its homepage.

Value

[ClusterFunctions].

[Package BatchJobs version 1.2 Index]