makeClusterFunctionsSSH {BatchJobs}	R Documentation

Create an SSH cluster to execute jobs.

Description

Worker nodes must share the same file system and be accessible by ssh without manually entering passwords (e.g. by ssh-agent or passwordless pubkey). Note that you can also use this function to parallelize on multiple cores on your local machine. But you still have to run an ssh server and provide passwordless access to localhost.

Usage

makeClusterFunctionsSSH(..., workers)

Arguments

...
[SSHWorker]
Worker objects, all created with makeSSHWorker.

workers
[list of SSHWorker]
Alternative way to pass workers.

Value

[ClusterFunctions].

See Also

ClusterFunctions makeSSHWorker

Examples

## Not run: 
##D 
##D # Assume you have three nodes larry, curley and moe. All have 6
##D # cpu cores. On curley and moe R is installed under
##D # "/opt/R/R-current" and on larry R is installed under
##D # "/usr/local/R/". larry should not be used extensively because
##D # somebody else wants to compute there as well.
##D # Then a call to 'makeClusterFunctionsSSH'
##D # might look like this:
##D 
##D cluster.functions = makeClusterFunctionsSSH(
##D   makeSSHWorker(nodename = "larry", rhome = "/usr/local/R", max.jobs = 2),
##D   makeSSHWorker(nodename = "curley", rhome = "/opt/R/R-current"),
##D   makeSSHWorker(nodename = "moe", rhome = "/opt/R/R-current"))
## End(Not run)

[Package BatchJobs version 1.3 Index]