makeSSHWorker {BatchJobs}	R Documentation

Create SSH worker for SSH cluster functions.

Description

Create SSH worker for SSH cluster functions.

Usage

makeSSHWorker(nodename, rhome = "", ncpus, max.jobs, max.load, nice,
  r.options = c("--no-save", "--no-restore", "--no-init-file",
  "--no-site-file"), script)

Arguments

nodename
[character(1)]
Host name of node.

rhome
[character(1)]
Path to R installation on worker. “” means R installation on the PATH is used, of course this implies that it must be on the PATH (also for non-interactive shells)! Default is “”.

ncpus
[integers(1)]
Number of VPUs of worker. Default means to query the worker via “/proc/cpuinfo”.

max.jobs
[integer(1)]
Maximal number of jobs that can run concurrently for the current registry. Default is ncpus.

max.load
[numeric(1)]
Load average (of the last 5 min) at which the worker is considered occupied, so that no job can be submitted. Default is ncpus-1.

nice
[integer(1)]
Process priority to run R with set via nice. Integers between -20 and 19 are allowed. If missing, processes are not nice'd and the system default applies (usually 0). Default is no niceing.

r.options
[character] Options for R and Rscript, one option per element of the vector, a la “–vanilla”. Default is c("--no-save", "--no-restore", "--no-init-file", "--no-site-file").

script
[character(1)]
Path to helper bash script which interacts with the worker. You really should not have to touch this, as this would imply that we have screwed up and published an incompatible version for your system. This option is only provided as a last resort for very experienced hackers. Note that the path has to be absolute. This is what is done in the package: https://github.com/tudo-r/BatchJobs/blob/master/inst/bin/linux-helper Default means to take it from package directory.

Value

[SSHWorker].

[Package BatchJobs version 1.3 Index]