makeSubmitJobResult {BatchJobs}	R Documentation

Create a SubmitJobResult object.

Description

Use this function in your implementation of makeClusterFunctions to create a return value for the submitJob function.

Usage

makeSubmitJobResult(status, batch.job.id, msg, ...)

Arguments

status
[integer(1)]
Launch status of job. 0 means success, codes bewteen 1 and 100 are temporary errors and any error greater than 100 is a permanent failure.

batch.job.id
[character(1)]
Unique id of this job on batch system. Note that this is not the usual job id used in BatchJobs! Must be globally unique so that the job can be terminated using just this information.

msg
[character(1)]
Optional error message in case status is not equal to 0. Default is “OK”, “TEMPERR”, “ERROR”, depending on status.

...
[any]
Currently unused.

Value

[SubmitJobResult]. A list, containing status, batch.job.id and msg.

[Package BatchJobs version 1.3 Index]