reduceResults {BatchJobs}	R Documentation

Reduce results from result directory.

Description

The following functions provide ways to reduce result files into either specific R objects (like vectors, lists, matrices or data.frames) or to arbitrarily aggregate them, which is a more general operation.

Usage

reduceResults(reg, ids, part = NA_character_, fun, init, impute.val, ...)

reduceResultsList(reg, ids, part = NA_character_, fun, ...,
  use.names = "ids", impute.val)

reduceResultsVector(reg, ids, part = NA_character_, fun, ...,
  use.names = "ids", impute.val)

reduceResultsMatrix(reg, ids, part = NA_character_, fun, ..., rows = TRUE,
  use.names = "ids", impute.val)

reduceResultsDataFrame(reg, ids, part = NA_character_, fun, ...,
  use.names = "ids", impute.val,
  strings.as.factors = default.stringsAsFactors())

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of selected jobs. Default is all jobs for which results are available.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

fun
[function]
For reduceResults, a function function(aggr, job, res, ...) to reduce things, for all others, a function function(job, res, ...) to select stuff. Here, job is the current job descriptor (see Job), result is the current result object and aggr are the so far aggregated results. When using reduceResults, your function should add the stuff you want to have from job and result to aggr and return that. When using the other reductions, you should select the stuff you want to have from job and result and return something that can be coerced to an element of the selected return data structure (reasonable conversion is tried internally). Default behavior for this argument is to return res, except for reduceResults where no default is available.

init
[ANY]
Initial element, as used in Reduce. Default is first result.

...
[any]
Additional arguments to fun.

use.names
[character(1)]
Name the results with job ids (“ids”), stored job names (“names”) or return a unnamed result (“none”). Default is ids.

impute.val
[any]
For reduceResults: If not missing, the value of impute.val is passed to function fun as argument res for jobs with missing results.
For the specialized reduction functions reduceResults[Type]: If not missing, impute.val is used as a replacement for the return value of fun on missing results.

rows
[logical(1)]
Should the selected vectors be used as rows (or columns) in the result matrix? Default is TRUE.

strings.as.factors
[logical(1)] Should all character columns in result be converted to factors? Default is default.stringsAsFactors().

Value

Aggregated results, return type depends on function. If ids is empty: reduceResults returns init (if available) or NULL, reduceResultsVector returns c(), reduceResultsList returns list(), reduceResultsMatrix returns matrix(0,0,0), reduceResultsDataFrame returns data.frame().

Examples

# generate results:
reg = makeRegistry(id = "BatchJobsExample", file.dir = tempfile(), seed = 123)

## Creating dir: /tmp/RtmpdPOzrX/file6eba1b9cea68
## Saving registry: /tmp/RtmpdPOzrX/file6eba1b9cea68/registry.RData

f = function(x) x^2
batchMap(reg, f, 1:5)

## Adding 5 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpdPOzrX/file6eba1b9cea68/conf.RData
## Submitting 5 chunks / 5 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 5 R scripts...
## Loading registry: /tmp/RtmpdPOzrX/file6eba1b9cea68/registry.RData
## Loading conf: 
## 2014-07-10 07:45:24: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-07-10 07:45:24
## Setting seed: 123
## Writing result file: /tmp/RtmpdPOzrX/file6eba1b9cea68/jobs/01/1-result.RData
## 2014-07-10 07:45:24: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba1b9cea68/registry.RData
## Loading conf: 
## 2014-07-10 07:45:24: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-07-10 07:45:24
## Setting seed: 124
## Writing result file: /tmp/RtmpdPOzrX/file6eba1b9cea68/jobs/02/2-result.RData
## 2014-07-10 07:45:24: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba1b9cea68/registry.RData
## Loading conf: 
## 2014-07-10 07:45:24: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-07-10 07:45:24
## Setting seed: 125
## Writing result file: /tmp/RtmpdPOzrX/file6eba1b9cea68/jobs/03/3-result.RData
## 2014-07-10 07:45:24: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba1b9cea68/registry.RData
## Loading conf: 
## 2014-07-10 07:45:24: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-07-10 07:45:24
## Setting seed: 126
## Writing result file: /tmp/RtmpdPOzrX/file6eba1b9cea68/jobs/04/4-result.RData
## 2014-07-10 07:45:24: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba1b9cea68/registry.RData
## Loading conf: 
## 2014-07-10 07:45:25: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-07-10 07:45:25
## Setting seed: 127
## Writing result file: /tmp/RtmpdPOzrX/file6eba1b9cea68/jobs/05/5-result.RData
## 2014-07-10 07:45:25: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Sending 5 submit messages...
## Might take some time, do not interrupt this!

# reduce results to a vector
reduceResultsVector(reg)

## Reducing 5 results...

##  1  2  3  4  5 
##  1  4  9 16 25

# reduce results to sum
reduceResults(reg, fun = function(aggr, job, res) aggr+res)

## Reducing

## [1] 55

reg = makeRegistry(id = "BatchJobsExample", file.dir = tempfile(), seed = 123)

## Creating dir: /tmp/RtmpdPOzrX/file6eba5f706c2f
## Saving registry: /tmp/RtmpdPOzrX/file6eba5f706c2f/registry.RData

f = function(x) list(a = x, b = as.character(2*x), c = x^2)
batchMap(reg, f, 1:5)

## Adding 5 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpdPOzrX/file6eba5f706c2f/conf.RData
## Submitting 5 chunks / 5 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 5 R scripts...
## Loading registry: /tmp/RtmpdPOzrX/file6eba5f706c2f/registry.RData
## Loading conf: 
## 2014-07-10 07:45:25: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-07-10 07:45:25
## Setting seed: 123
## Writing result file: /tmp/RtmpdPOzrX/file6eba5f706c2f/jobs/01/1-result.RData
## 2014-07-10 07:45:25: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba5f706c2f/registry.RData
## Loading conf: 
## 2014-07-10 07:45:25: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-07-10 07:45:25
## Setting seed: 124
## Writing result file: /tmp/RtmpdPOzrX/file6eba5f706c2f/jobs/02/2-result.RData
## 2014-07-10 07:45:25: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba5f706c2f/registry.RData
## Loading conf: 
## 2014-07-10 07:45:25: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-07-10 07:45:25
## Setting seed: 125
## Writing result file: /tmp/RtmpdPOzrX/file6eba5f706c2f/jobs/03/3-result.RData
## 2014-07-10 07:45:25: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba5f706c2f/registry.RData
## Loading conf: 
## 2014-07-10 07:45:25: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-07-10 07:45:25
## Setting seed: 126
## Writing result file: /tmp/RtmpdPOzrX/file6eba5f706c2f/jobs/04/4-result.RData
## 2014-07-10 07:45:25: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba5f706c2f/registry.RData
## Loading conf: 
## 2014-07-10 07:45:25: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-07-10 07:45:25
## Setting seed: 127
## Writing result file: /tmp/RtmpdPOzrX/file6eba5f706c2f/jobs/05/5-result.RData
## 2014-07-10 07:45:25: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Sending 5 submit messages...
## Might take some time, do not interrupt this!

# reduce results to a vector
reduceResultsVector(reg, fun = function(job, res) res$a)

## Reducing 5 results...

## 1 2 3 4 5 
## 1 2 3 4 5

reduceResultsVector(reg, fun = function(job, res) res$b)

## Reducing 5 results...

##    1    2    3    4    5 
##  "2"  "4"  "6"  "8" "10"

# reduce results to a list
reduceResultsList(reg)

## Reducing 5 results...

## $`1`
## $`1`$a
## [1] 1
## 
## $`1`$b
## [1] "2"
## 
## $`1`$c
## [1] 1
## 
## 
## $`2`
## $`2`$a
## [1] 2
## 
## $`2`$b
## [1] "4"
## 
## $`2`$c
## [1] 4
## 
## 
## $`3`
## $`3`$a
## [1] 3
## 
## $`3`$b
## [1] "6"
## 
## $`3`$c
## [1] 9
## 
## 
## $`4`
## $`4`$a
## [1] 4
## 
## $`4`$b
## [1] "8"
## 
## $`4`$c
## [1] 16
## 
## 
## $`5`
## $`5`$a
## [1] 5
## 
## $`5`$b
## [1] "10"
## 
## $`5`$c
## [1] 25

# reduce results to a matrix
reduceResultsMatrix(reg, fun = function(job, res) res[c(1,3)])

## Reducing 5 results...

##   a  c
## 1 1  1
## 2 2  4
## 3 3  9
## 4 4 16
## 5 5 25

reduceResultsMatrix(reg, fun = function(job, res) c(foo = res$a, bar = res$c), rows = TRUE)

## Reducing 5 results...

##   foo bar
## 1   1   1
## 2   2   4
## 3   3   9
## 4   4  16
## 5   5  25

reduceResultsMatrix(reg, fun = function(job, res) c(foo = res$a, bar = res$c), rows = FALSE)

## Reducing 5 results...

##     1 2 3  4  5
## foo 1 2 3  4  5
## bar 1 4 9 16 25

# reduce results to a data.frame
print(str(reduceResultsDataFrame(reg)))

## Reducing 5 results...

## 'data.frame':    5 obs. of  3 variables:
##  $ a: int  1 2 3 4 5
##  $ b: Factor w/ 5 levels "10","2","4","6",..: 2 3 4 5 1
##  $ c: num  1 4 9 16 25
## NULL

# reduce results to a sum
reduceResults(reg, fun = function(aggr, job, res) aggr+res$a, init = 0)

## Reducing

## [1] 15

[Package BatchJobs version 1.3 Index]