resetJobs {BatchJobs}	R Documentation

Reset computational state of jobs.

Description

Reset state of jobs in the database. Useful under two circumstances: Either to re-submit them because of changes in e.g. external data or to resolve rare issues when jobs are killed in an unfortunate state and therefore blocking your registry.

Usage

resetJobs(reg, ids, force = FALSE)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs to kill. Default is none.

force
[logical(1)]
Reset jobs without checking whether they are currently running. READ THE DETAILS SECTION! Default is FALSE.

Details

The function internally lists all jobs on the batch system and if those include some of the jobs you want to reset, it informs you to kill them first by raising an exception. If you really know what you are doing, you may set force to TRUE to omit this sanity check. Note that this is a dangerous operation to perform which may harm the database integrity. In this case you HAVE to make externally sure that none of the jobs you want to reset are still running.

Value

Vector of reseted job ids.

[Package BatchJobs version 1.3 Index]