setJobFunction {BatchJobs}	R Documentation

Sets the job function for already existing jobs.

Description

Use this function only as last measure when there is a bug in a part of your job function and you have already computed a large number of (unaffected) results. This function allows you to fix the error and to associate the jobs with the corrected function.

Usage

  setJobFunction(reg, ids, fun, more.args = list(),
    reset = TRUE, force = FALSE)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs. Default is all jobs.

fun
[function]
Replacement function.

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

reset
[logical(1)]
Reset job status via resetJobs. Default is TRUE.

force
[logical(1)]
See resetJobs. Default is FALSE.

Details

Note that by default the computational state of the affected jobs is also reset.

Value

Nothing.

[Package BatchJobs version 1.2 Index]