showClusterStatus {BatchJobs}	R Documentation

Show information about available computational resources on cluster.

Description

Currently only supported for multicore and SSH mode. Displays: Name of node, current load, number of running R processes, number of R processes with more than 50 number of BatchJobs jobs running. The latter counts either jobs belonging to reg or all BatchJobs jobs if reg was not passed.

Usage

  showClusterStatus(reg)

Arguments

reg
[Registry]
Registry. Must not be passed and this is the default.

Value

[data.frame]. Returns the displayed table invisibly.

[Package BatchJobs version 1.2 Index]