showLog {BatchJobs}	R Documentation

Display the contents of a log file.

Description

Display the contents of a log file, useful in case of errors.

Usage

showLog(reg, id, pager = getOption("pager"))

Arguments

reg
[Registry]
Registry.

id
[integer(1)]
Id of selected job. Default is first id in registry.

pager
[any]
Pager to use to display the log. Defaults to getOption("pager"). This option is passed to file.show and is highly OS dependant and GUI dependant. If either R's pager or the environment variable “PAGER” is set to “less” or “vim”, the correct part of the log file will be shown. Otherwise you find information about the correct part in the beginning of the displayed file.

Details

Note this rare special case: When you use chunking, submit some jobs, some jobs fail, then you resubmit these jobs again in different chunks, the log files will contain the log of the old, failed job as well. showLog tries to jump to the correct part of the new log file with a supported pager.

Value

[character(1)]. Invisibly returns path to log file.

See Also

grepLogs, getErrorMessages

[Package BatchJobs version 1.3 Index]