submitJobs {BatchJobs}	R Documentation

Submit jobs or chunks of jobs to batch system via cluster function.

Description

If the internal submit cluster function completes successfully, the retries counter is set back to 0 and the next job or chunk is submitted. If the internal submit cluster function returns a fatal error, the submit process is completely stopped and an exception is thrown. If the internal submit cluster function returns a temporary error, the submit process waits for a certain time, which is determined by calling the user-defined wait-function with the current retries counter, the counter is increased by 1 and the same job is submitted again. If max.retries is reached the function simply terminates.

Potential temporary submit warnings and errors are logged inside your file directory in the file “submit.log”. To keep track you can use tail -f [file.dir]/submit.log in another terminal.

Usage

submitJobs(reg, ids, resources = list(), wait, max.retries = 10L,
  chunks.as.arrayjobs = FALSE, job.delay = FALSE)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Vector for job id or list of vectors of chunked job ids. Only corresponding jobs are submitted. Chunked jobs will get executed sequentially as a single job for the scheduler. Default is all jobs which were not yet submitted to the batch system.

resources
[list]
Required resources for all batch jobs. The elements of this list (e.g. something like “walltime” or “nodes” are defined by your template job file. Defaults can be specified in your config file. Default is empty list.

wait
[function(retries)]
Function that defines how many seconds should be waited in case of a temporary error. Default is exponential back-off with 10*2^retries.

max.retries
[integer(1)]
Number of times to submit one job again in case of a temporary error (like filled queues). Each time wait is called to wait a certain number of seconds. Default is 10 times.

chunks.as.arrayjobs
[logical(1)]
If ids are passed as a list of chunked job ids, execute jobs in a chunk as array jobs. Note that your scheduler and your template must be adjusted to use this option. Default is FALSE.

job.delay
[function(n, i) or logical(1)]
Function that defines how many seconds a job should be delayed before it starts. This is an expert option and only necessary to change when you want submit extremely many jobs. We then delay the jobs a bit to write the submit messages as early as possible to avoid writer starvation. n is the number of jobs and i the number of the ith job. The default function used with job.delay set to TRUE is no delay for 100 jobs or less and otherwise runif(1, 0.1*n, 0.2*n). If set to FALSE (the default) delaying jobs is disabled.

Value

[integer]. Vector of submitted job ids.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)

## Creating dir: /tmp/Rtmp14EkvD/file3d4b2f0317ec
## Saving registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData

f <- function(x) x^2
batchMap(reg, f, 1:10)

## Adding 10 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Submitting 10 chunks / 10 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 10 R scripts...
## 2014-02-07 07:55:50: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-02-07 07:55:50
## Setting seed: 123
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/01/1-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:51: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-02-07 07:55:51
## Setting seed: 124
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/02/2-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:51: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-02-07 07:55:51
## Setting seed: 125
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/03/3-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:51: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-02-07 07:55:51
## Setting seed: 126
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/04/4-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:51: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-02-07 07:55:51
## Setting seed: 127
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/05/5-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:51: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-02-07 07:55:51
## Setting seed: 128
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/06/6-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:51: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-02-07 07:55:51
## Setting seed: 129
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/07/7-result.RData
## 2014-02-07 07:55:51: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:52: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 130
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/08/8-result.RData
## 2014-02-07 07:55:52: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:52: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 131
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/09/9-result.RData
## 2014-02-07 07:55:52: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:52: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 132
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/10/10-result.RData
## 2014-02-07 07:55:52: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |+++++                                            |  10% (00:00:09)
SubmitJobs |++++++++++                                       |  20% (00:00:04)
SubmitJobs |+++++++++++++++                                  |  30% (00:00:02)
SubmitJobs |++++++++++++++++++++                             |  40% (00:00:01)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++                    |  60% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++               |  70% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++          |  80% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 10 submit messages...
## Might take some time, do not interrupt this!

# Submit the 10 jobs again, now randomized into 2 chunks:
chunked = chunk(getJobIds(reg), n.chunks=2, shuffle=TRUE)
submitJobs(reg, chunked)

## Saving conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Submitting 2 chunks / 10 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 2 R scripts...
## 2014-02-07 07:55:52: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## 2014-02-07 07:55:52: Waiting 557 secs between msg flushes.
## ########## Executing jid=1 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 123
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/01/1-result.RData
## ########## Executing jid=5 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 127
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/05/5-result.RData
## ########## Executing jid=6 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 128
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/06/6-result.RData
## ########## Executing jid=8 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 130
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/08/8-result.RData
## ########## Executing jid=10 ##########
## Timestamp: 2014-02-07 07:55:52
## Setting seed: 132
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/10/10-result.RData
## 2014-02-07 07:55:52: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 07:55:53: Starting job on node bischl-x1.
## Loading registry: /tmp/Rtmp14EkvD/file3d4b2f0317ec/registry.RData
## Loading conf: /tmp/Rtmp14EkvD/file3d4b2f0317ec/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## 2014-02-07 07:55:53: Waiting 382 secs between msg flushes.
## ########## Executing jid=2 ##########
## Timestamp: 2014-02-07 07:55:53
## Setting seed: 124
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/02/2-result.RData
## ########## Executing jid=3 ##########
## Timestamp: 2014-02-07 07:55:53
## Setting seed: 125
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/03/3-result.RData
## ########## Executing jid=4 ##########
## Timestamp: 2014-02-07 07:55:53
## Setting seed: 126
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/04/4-result.RData
## ########## Executing jid=7 ##########
## Timestamp: 2014-02-07 07:55:53
## Setting seed: 129
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/07/7-result.RData
## ########## Executing jid=9 ##########
## Timestamp: 2014-02-07 07:55:53
## Setting seed: 131
## Writing result file: /tmp/Rtmp14EkvD/file3d4b2f0317ec/jobs/09/9-result.RData
## 2014-02-07 07:55:53: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 2 submit messages...
## Might take some time, do not interrupt this!

[Package BatchJobs version 1.2 Index]