sweepRegistry {BatchJobs}	R Documentation

Sweep obsolete files from the file system

Description

Removes R scripts, log files, resource informations and temporarily stored configuration files from the registry's file directory. Assuming all your jobs completed successfully, none of these are needed for further work. This operation potentially releases quite a lot of disk space, depending on the number of your jobs. BUT A HUGE WORD OF WARNING: IF you later notice something strange and need to determine the reason for it, you are at a huge disadvantage. Only do this at your own risk and when you are sure that you have successfully completed a project and only want to archive your produced experiments and results.

Usage

  sweepRegistry(reg, sweep = c("scripts", "conf"))

Arguments

reg
[Registry]
Registry.

sweep
[character]
Possible choices: Temporary R scripts of jobs, really not needed for anything else then execution (“scripts”), log file of jobs, think about whether you later want to inspect them (“logs”), BatchJobs configuration files which are temporarily stored on submit, really not needed for anything else then execution (“conf”), resource lists of submitJobs which are temporarily stored on submit, think about whether you later want to inspect them (“resources”), Default is c("scripts", "conf").

Value

[logical]. Invisibly returns TRUE on success and FALSE if some files could not be removed.

[Package BatchJobs version 1.2 Index]