waitForJobs {BatchJobs}	R Documentation

Wait for termination of jobs on the batch system.

Description

Waits for termination of jobs while displaying a progress bar containing summarizing informations of the jobs. The following abbreviations are used in the progress bar: “S” for number of jobs on system, “D” for number of jobs successfully terminated, “E” for number ofjobs terminated with an R exception and “R” for number of jobs currently running on the system.

Usage

waitForJobs(reg, ids, sleep = 10, timeout = 604800, stop.on.error = FALSE)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Vector of job ids. Default is all submitted jobs not yet terminated.

sleep
[numeric(1)]
Seconds to sleep between status updates. Default is 10.

timeout
[numeric(1)]
After waiting timeout seconds, show a message and return FALSE. This argument may be required on some systems where, e.g., expired jobs or jobs on hold are problematic to detect. If you don't want a timeout, set this to Inf. Default is 604800 (one week).

stop.on.error
[logical(1)]
Immediately return if a job terminates with an error? Default is FALSE.

Value

[logical(1)]. Returns TRUE if all jobs terminated successfully and FALSE if either an error occurred or the timeout is reached.

[Package BatchJobs version 1.3 Index]