batchExpandGrid {BatchJobs}	R Documentation

Map function over all combinations.

Description

Maps an n-ary-function over a list of all combinations which are given by some vectors. Internally expand.grid is used to compute the combinations, then batchMap is called.

Usage

batchExpandGrid(reg, fun, ..., more.args = list())

Arguments

reg
[Registry]
Empty Registry that will store jobs for the mapping.

fun
[function]
Function to map over the combinations.

...
[any]
Vectors that are used to compute all combinations. If the arguments are named, these names are used to bind to arguments of fun.

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

Value

[data.frame]. Expanded grid of combinations produced by expand.grid.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)

## Creating dir: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4
## Saving registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData

f <- function(x, y, z) x * y  + z
batchExpandGrid(reg, f, x=1:2, y=1:3, more.args=list(z=10))

## Adding 6 jobs to DB.

##   x y
## 1 1 1
## 2 2 1
## 3 1 2
## 4 2 2
## 5 1 3
## 6 2 3

submitJobs(reg)

## Saving conf: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/conf.RData
## Submitting 6 chunks / 6 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 6 R scripts...
## Loading registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData
## Loading conf: 
## 2014-05-15 23:35:04: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-05-15 23:35:04
## Setting seed: 123
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/jobs/01/1-result.RData
## 2014-05-15 23:35:04: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData
## Loading conf: 
## 2014-05-15 23:35:04: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-05-15 23:35:04
## Setting seed: 124
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/jobs/02/2-result.RData
## 2014-05-15 23:35:04: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData
## Loading conf: 
## 2014-05-15 23:35:04: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-05-15 23:35:04
## Setting seed: 125
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/jobs/03/3-result.RData
## 2014-05-15 23:35:04: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData
## Loading conf: 
## 2014-05-15 23:35:05: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-05-15 23:35:05
## Setting seed: 126
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/jobs/04/4-result.RData
## 2014-05-15 23:35:05: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData
## Loading conf: 
## 2014-05-15 23:35:05: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-05-15 23:35:05
## Setting seed: 127
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/jobs/05/5-result.RData
## 2014-05-15 23:35:05: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/registry.RData
## Loading conf: 
## 2014-05-15 23:35:05: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-05-15 23:35:05
## Setting seed: 128
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmp4xHpvc/file23e856cc45d4/jobs/06/6-result.RData
## 2014-05-15 23:35:05: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++++++++                                         |  17% (00:00:00)
SubmitJobs |++++++++++++++++                                 |  33% (00:00:00)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++                |  67% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++        |  83% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 6 submit messages...
## Might take some time, do not interrupt this!

reduceResultsMatrix(reg, fun=function(job, res) cbind(job$pars, res))

## Reducing 6 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |++++++++                                      |  17% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++                       |  50% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++        |  83% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

##   [,1] [,2] [,3] [,4]
## 1    1    1   11   11
## 2    2    1   12   12
## 3    1    2   12   12
## 4    2    2   14   14
## 5    1    3   13   13
## 6    2    3   16   16

[Package BatchJobs version 1.3 Index]