batchReduce {BatchJobs}	R Documentation

Reduces via a binary function over a list adding jobs to a registry.

Description

Each jobs reduces a certain number of elements on one slave. You can then submit these jobs to the batch system.

Usage

batchReduce(reg, fun, xs, init, block.size, more.args = list())

Arguments

reg
[Registry]
Empty Registry.

fun
[function(aggr, x, ...)]
Function to reduce xs with.

xs
[vector]
Vector to reduce.

init
[any]
Initial object for reducing.

block.size
[integer(1)]
Number of elements of xs reduced in one job.

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

Value

Vector of type integer with job ids.

Examples

# define function to reduce on slave, we want to sum a vector
f = function(aggr, x) aggr + x
reg = makeRegistry(id = "BatchJobsExample", file.dir = tempfile(), seed = 123)

## Creating dir: /tmp/RtmpdPOzrX/file6eba4ecbfa85
## Saving registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData

# sum 20 numbers on each slave process, i.e. 5 jobs
batchReduce(reg, fun = f, 1:100, init = 0, block.size = 5)

## Adding 20 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpdPOzrX/file6eba4ecbfa85/conf.RData
## Submitting 20 chunks / 20 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 20 R scripts...
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:19: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 123
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/01/1-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 124
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/02/2-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 125
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/03/3-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 126
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/04/4-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 127
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/05/5-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 128
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/06/6-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 129
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/07/7-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 130
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/08/8-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 131
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/09/9-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 132
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/10/10-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=11 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 133
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/11/11-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=12 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 134
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/12/12-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=13 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 135
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/13/13-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=14 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 136
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/14/14-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=15 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 137
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/15/15-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:20: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=16 ##########
## Timestamp: 2014-07-10 07:45:20
## Setting seed: 138
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/16/16-result.RData
## 2014-07-10 07:45:20: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:21: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=17 ##########
## Timestamp: 2014-07-10 07:45:21
## Setting seed: 139
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/17/17-result.RData
## 2014-07-10 07:45:21: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:21: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=18 ##########
## Timestamp: 2014-07-10 07:45:21
## Setting seed: 140
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/18/18-result.RData
## 2014-07-10 07:45:21: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:21: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=19 ##########
## Timestamp: 2014-07-10 07:45:21
## Setting seed: 141
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/19/19-result.RData
## 2014-07-10 07:45:21: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Loading registry: /tmp/RtmpdPOzrX/file6eba4ecbfa85/registry.RData
## Loading conf: 
## 2014-07-10 07:45:21: Starting job on node bischl-x1.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=20 ##########
## Timestamp: 2014-07-10 07:45:21
## Setting seed: 142
## Writing result file: /tmp/RtmpdPOzrX/file6eba4ecbfa85/jobs/20/20-result.RData
## 2014-07-10 07:45:21: All done.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## Sending 20 submit messages...
## Might take some time, do not interrupt this!

# now reduce one final time on master
reduceResults(reg, fun = function(aggr,job,res) f(aggr, res))

## Reducing

## [1] 5050

[Package BatchJobs version 1.3 Index]