batchReduceResults {BatchJobs}	R Documentation

Reduces results via a binary function and adds jobs for this to a registry.

Description

Each jobs reduces a certain number of results on one slave. You can then submit these jobs to the batch system. Later, you can do a final reduction with reduceResults on the master.

Usage

batchReduceResults(reg, reg2, fun, ids, part = NA_character_, init,
  block.size, more.args = list())

Arguments

reg
[Registry]
Registry whose results should be reduced by fun.

reg2
[Registry]
Empty registry that should store the job for the mapping.

fun
[function(aggr, job, res, ...)]
Function to reduce results with.

ids
[integer]
Ids of jobs whose results should be reduced with fun. Default is all jobs.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

init
[any]
Initial object for reducing.

block.size
[integer(1)]
Number of results reduced in one job.

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

Value

Vector of type integer with job ids.

Examples

# generating example results:
reg1 <- makeRegistry(id="BatchJobsExample1", file.dir=tempfile(), seed=123)
f <- function(x) x^2
batchMap(reg1, f, 1:20)
submitJobs(reg1)

## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 123
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/01/1-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 124
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/02/2-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 125
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/03/3-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 126
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/04/4-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 127
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/05/5-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 128
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/06/6-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 129
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/07/7-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 130
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/08/8-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 131
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/09/9-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 132
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/10/10-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:37: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=11 ##########
## Timestamp: 2014-05-15 23:44:37
## Setting seed: 133
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/11/11-result.RData
## 2014-05-15 23:44:37: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=12 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 134
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/12/12-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=13 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 135
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/13/13-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=14 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 136
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/14/14-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=15 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 137
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/15/15-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=16 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 138
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/16/16-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=17 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 139
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/17/17-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=18 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 140
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/18/18-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=19 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 141
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/19/19-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=20 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 142
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc4dbc6161/jobs/20/20-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:

# define function to reduce on slave, we want to sum the squares
myreduce <- function(aggr, job, res) aggr + res

# sum 5 results on each slave process, i.e. 4 jobs
reg2 <- makeRegistry(id="BatchJobsExample2", file.dir=tempfile(), seed=123)
batchReduceResults(reg1, reg2, fun=myreduce, init=0, block.size=5)
submitJobs(reg2)

## 2014-05-15 23:44:38: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-05-15 23:44:38
## Setting seed: 123
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfca9d4f26/jobs/01/1-result.RData
## 2014-05-15 23:44:38: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 124
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfca9d4f26/jobs/02/2-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 125
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfca9d4f26/jobs/03/3-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 126
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfca9d4f26/jobs/04/4-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:

# now reduce one final time on master
reduceResults(reg2, fun=myreduce)

## [1] 2870

[Package BatchJobs version 1.3 Index]