batchReduceResults {BatchJobs}	R Documentation

Reduces results via a binary function and adds jobs for this to a registry.

Description

Each jobs reduces a certain number of results on one slave. You can then submit these jobs to the batch system. Later, you can do a final reduction with reduceResults on the master.

Usage

  batchReduceResults(reg, reg2, fun, ids,
    part = NA_character_, init, block.size,
    more.args = list())

Arguments

reg
[Registry]
Registry whose results should be reduced by fun.

reg2
[Registry]
Empty registry that should store the job for the mapping.

fun
[function(aggr, job, res, ...)]
Function to reduce results with.

ids
[integer]
Ids of jobs whose results should be reduced with fun. Default is all jobs.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

init
[any]
Initial object for reducing.

block.size
[integer(1)]
Number of results reduced in one job.

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

Value

Vector of type integer with job ids.

Examples

# generating example results:
reg1 <- makeRegistry(id="BatchJobsExample1", file.dir=tempfile(), seed=123)

## Creating dir: /tmp/RtmpTp3YLN/file415571d2b19a
## Saving registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData

f <- function(x) x^2
batchMap(reg1, f, 1:20)

## Adding 20 jobs to DB.

submitJobs(reg1)

## Saving conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Submitting 20 chunks / 20 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 20 R scripts...
## 2013-11-26 11:41:53: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2013-11-26 11:41:53
## Setting seed: 123
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/01/1-result.RData
## 2013-11-26 11:41:53: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:53: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2013-11-26 11:41:53
## Setting seed: 124
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/02/2-result.RData
## 2013-11-26 11:41:53: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 125
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/03/3-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 126
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/04/4-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 127
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/05/5-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 128
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/06/6-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 129
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/07/7-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 130
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/08/8-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:54: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2013-11-26 11:41:54
## Setting seed: 131
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/09/9-result.RData
## 2013-11-26 11:41:54: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 132
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/10/10-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=11 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 133
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/11/11-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=12 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 134
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/12/12-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=13 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 135
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/13/13-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=14 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 136
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/14/14-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=15 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 137
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/15/15-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=16 ##########
## Timestamp: 2013-11-26 11:41:55
## Setting seed: 138
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/16/16-result.RData
## 2013-11-26 11:41:55: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:55: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=17 ##########
## Timestamp: 2013-11-26 11:41:56
## Setting seed: 139
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/17/17-result.RData
## 2013-11-26 11:41:56: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:56: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=18 ##########
## Timestamp: 2013-11-26 11:41:56
## Setting seed: 140
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/18/18-result.RData
## 2013-11-26 11:41:56: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:56: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=19 ##########
## Timestamp: 2013-11-26 11:41:56
## Setting seed: 141
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/19/19-result.RData
## 2013-11-26 11:41:56: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:56: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file415571d2b19a/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file415571d2b19a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=20 ##########
## Timestamp: 2013-11-26 11:41:56
## Setting seed: 142
## Writing result file: /tmp/RtmpTp3YLN/file415571d2b19a/jobs/20/20-result.RData
## 2013-11-26 11:41:56: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++                                               |   5% (00:00:00)
SubmitJobs |+++++                                            |  10% (00:00:09)
SubmitJobs |+++++++                                          |  15% (00:00:05)
SubmitJobs |++++++++++                                       |  20% (00:00:04)
SubmitJobs |++++++++++++                                     |  25% (00:00:03)
SubmitJobs |+++++++++++++++                                  |  30% (00:00:02)
SubmitJobs |+++++++++++++++++                                |  35% (00:00:01)
SubmitJobs |++++++++++++++++++++                             |  40% (00:00:01)
SubmitJobs |++++++++++++++++++++++                           |  45% (00:00:02)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++                      |  55% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++                    |  60% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++                 |  65% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++++               |  70% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++            |  75% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++          |  80% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++       |  85% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++  |  95% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 20 submit messages...
## Might take some time, do not interrupt this!

# define function to reduce on slave, we want to sum the squares
myreduce <- function(aggr, job, res) aggr + res

# sum 5 results on each slave process, i.e. 4 jobs
reg2 <- makeRegistry(id="BatchJobsExample2", file.dir=tempfile(), seed=123)

## Creating dir: /tmp/RtmpTp3YLN/file41551f7da544
## Saving registry: /tmp/RtmpTp3YLN/file41551f7da544/registry.RData

batchReduceResults(reg1, reg2, fun=myreduce, init=0, block.size=5)

## Saving registry: /tmp/RtmpTp3YLN/file41551f7da544/registry.RData
## Adding 4 jobs to DB.

submitJobs(reg2)

## Saving conf: /tmp/RtmpTp3YLN/file41551f7da544/conf.RData
## Submitting 4 chunks / 4 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 4 R scripts...
## 2013-11-26 11:41:56: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41551f7da544/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41551f7da544/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2013-11-26 11:41:56
## Setting seed: 123
## Writing result file: /tmp/RtmpTp3YLN/file41551f7da544/jobs/01/1-result.RData
## 2013-11-26 11:41:56: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:56: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41551f7da544/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41551f7da544/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2013-11-26 11:41:56
## Setting seed: 124
## Writing result file: /tmp/RtmpTp3YLN/file41551f7da544/jobs/02/2-result.RData
## 2013-11-26 11:41:56: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:57: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41551f7da544/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41551f7da544/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2013-11-26 11:41:57
## Setting seed: 125
## Writing result file: /tmp/RtmpTp3YLN/file41551f7da544/jobs/03/3-result.RData
## 2013-11-26 11:41:57: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:57: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41551f7da544/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41551f7da544/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2013-11-26 11:41:57
## Setting seed: 126
## Writing result file: /tmp/RtmpTp3YLN/file41551f7da544/jobs/04/4-result.RData
## 2013-11-26 11:41:57: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++++++++++++                                     |  25% (00:00:00)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++++            |  75% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 4 submit messages...
## Might take some time, do not interrupt this!

# now reduce one final time on master
reduceResults(reg2, fun=myreduce)

## Reducing 4 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |++++++++++++                                  |  25% (00:00:00)
reduceResults |+++++++++++++++++++++++                       |  50% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++            |  75% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## [1] 2870

[Package BatchJobs version 1.2 Index]