filterResults {BatchJobs}	R Documentation

Find all results where a specific condition is true.

Description

Find all results where a specific condition is true.

Usage

filterResults(reg, ids, fun, ...)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs whose results you want to test for the condition. Default is all jobs for which results are available.

fun
[fun(job, res)]
Predicate function that returns TRUE or FALSE.

...
[any]
Additional arguments to fun.

Value

[integer]. Ids of jobs where fun(job, result) returns TRUE.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)
f <- function(x) x^2
batchMap(reg, f, 1:10)
submitJobs(reg)

## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 123
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/01/1-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 124
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/02/2-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 125
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/03/3-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 126
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/04/4-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 127
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/05/5-result.RData
## 2014-05-15 23:44:39: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:39: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-05-15 23:44:39
## Setting seed: 128
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/06/6-result.RData
## 2014-05-15 23:44:40: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:40: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-05-15 23:44:40
## Setting seed: 129
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/07/7-result.RData
## 2014-05-15 23:44:40: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:40: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-05-15 23:44:40
## Setting seed: 130
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/08/8-result.RData
## 2014-05-15 23:44:40: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:40: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-05-15 23:44:40
## Setting seed: 131
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/09/9-result.RData
## 2014-05-15 23:44:40: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:40: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-05-15 23:44:40
## Setting seed: 132
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfcf051526/jobs/10/10-result.RData
## 2014-05-15 23:44:40: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:

# which square numbers are even:
filterResults(reg, fun=function(job, res) res %% 2 == 0)

## [1]  2  4  6  8 10

[Package BatchJobs version 1.3 Index]