filterResults {BatchJobs}	R Documentation

Find all results where a specific condition is true.

Description

Find all results where a specific condition is true.

Usage

  filterResults(reg, ids, fun, ...)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of jobs whose results you want to test for the condition. Default is all jobs for which results are available.

fun
[fun(job, res)]
Predicate function that returns TRUE or FALSE.

...
[any]
Additional arguments to fun.

Value

[integer]. Ids of jobs where fun(job, result) returns TRUE.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)

## Creating dir: /tmp/RtmpTp3YLN/file41554994850
## Saving registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData

f <- function(x) x^2
batchMap(reg, f, 1:10)

## Adding 10 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Submitting 10 chunks / 10 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 10 R scripts...
## 2013-11-26 11:41:57: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2013-11-26 11:41:57
## Setting seed: 123
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/01/1-result.RData
## 2013-11-26 11:41:57: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:57: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 124
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/02/2-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 125
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/03/3-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 126
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/04/4-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 127
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/05/5-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 128
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/06/6-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 129
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/07/7-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 130
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/08/8-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:58: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2013-11-26 11:41:58
## Setting seed: 131
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/09/9-result.RData
## 2013-11-26 11:41:58: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:41:59: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file41554994850/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file41554994850/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2013-11-26 11:41:59
## Setting seed: 132
## Writing result file: /tmp/RtmpTp3YLN/file41554994850/jobs/10/10-result.RData
## 2013-11-26 11:41:59: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |+++++                                            |  10% (00:00:00)
SubmitJobs |++++++++++                                       |  20% (00:00:04)
SubmitJobs |+++++++++++++++                                  |  30% (00:00:02)
SubmitJobs |++++++++++++++++++++                             |  40% (00:00:01)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++                    |  60% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++               |  70% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++          |  80% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 10 submit messages...
## Might take some time, do not interrupt this!

# which square numbers are even:
filterResults(reg, fun=function(job, res) res %% 2 == 0)

## [1]  2  4  6  8 10

[Package BatchJobs version 1.2 Index]