showStatus {BatchJobs}	R Documentation

Show status information about jobs.

Description

E.g.: How many there are, how many are done, any errors, etc.

Usage

  showStatus(reg, ids, run.and.exp = TRUE, errors = 10L)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of selected jobs. Default is all jobs.

errors
[integer(1)]
How many of the error messages should be displayed if any errors occured in the jobs? Default is 10.

run.and.exp
[logical(1)]
Show running and expired jobs? Requires to list the job on the batch system. If not possible, because that cluster function is not avilable, this option is ignored anyway. Default is TRUE.

Value

[list] List of absolute job numbers as printed by showStatus. Returned invisibly.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)

## Creating dir: /tmp/RtmpTp3YLN/file4155264bf21d
## Saving registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData

f <- function(x) x^2
batchMap(reg, f, 1:10)

## Adding 10 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Submitting 10 chunks / 10 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 10 R scripts...
## 2013-11-26 11:42:03: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2013-11-26 11:42:03
## Setting seed: 123
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/01/1-result.RData
## 2013-11-26 11:42:03: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:03: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2013-11-26 11:42:03
## Setting seed: 124
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/02/2-result.RData
## 2013-11-26 11:42:03: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:03: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2013-11-26 11:42:03
## Setting seed: 125
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/03/3-result.RData
## 2013-11-26 11:42:03: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:03: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2013-11-26 11:42:03
## Setting seed: 126
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/04/4-result.RData
## 2013-11-26 11:42:03: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:03: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2013-11-26 11:42:03
## Setting seed: 127
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/05/5-result.RData
## 2013-11-26 11:42:03: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:03: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2013-11-26 11:42:03
## Setting seed: 128
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/06/6-result.RData
## 2013-11-26 11:42:03: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:04: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2013-11-26 11:42:04
## Setting seed: 129
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/07/7-result.RData
## 2013-11-26 11:42:04: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:04: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2013-11-26 11:42:04
## Setting seed: 130
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/08/8-result.RData
## 2013-11-26 11:42:04: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:04: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2013-11-26 11:42:04
## Setting seed: 131
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/09/9-result.RData
## 2013-11-26 11:42:04: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2013-11-26 11:42:04: Starting job on node Bischl.
## Loading registry: /tmp/RtmpTp3YLN/file4155264bf21d/registry.RData
## Loading conf: /tmp/RtmpTp3YLN/file4155264bf21d/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2013-11-26 11:42:04
## Setting seed: 132
## Writing result file: /tmp/RtmpTp3YLN/file4155264bf21d/jobs/10/10-result.RData
## 2013-11-26 11:42:04: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |+++++                                            |  10% (00:00:00)
SubmitJobs |++++++++++                                       |  20% (00:00:00)
SubmitJobs |+++++++++++++++                                  |  30% (00:00:00)
SubmitJobs |++++++++++++++++++++                             |  40% (00:00:00)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++                    |  60% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++               |  70% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++          |  80% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 10 submit messages...
## Might take some time, do not interrupt this!

# after a few moments:
showStatus(reg)

## Status for 10 jobs at 2013-11-26 11:42:04
## Submitted: 10 (100.00%)
## Started:   10 (100.00%)
## Running:    0 (  0.00%)
## Done:      10 (100.00%)
## Errors:     0 (  0.00%)
## Expired:    0 (  0.00%)
## Time: min=0.00s avg=0.00s max=0.00s

# should show 10 submitted jobs, which are all done.

[Package BatchJobs version 1.2 Index]