showStatus {BatchJobs}	R Documentation

Show status information about jobs.

Description

E.g.: How many there are, how many are done, any errors, etc.

Usage

showStatus(reg, ids, run.and.exp = TRUE, errors = 10L)

Arguments

reg
[Registry]
Registry.

ids
[integer]
Ids of selected jobs. Default is all jobs.

errors
[integer(1)]
How many of the error messages should be displayed if any errors occured in the jobs? Default is 10.

run.and.exp
[logical(1)]
Show running and expired jobs? Requires to list the job on the batch system. If not possible, because that cluster function is not avilable, this option is ignored anyway. Default is TRUE.

Value

[list] List of absolute job numbers as printed by showStatus. Returned invisibly.

Examples

reg <- makeRegistry(id="BatchJobsExample", file.dir=tempfile(), seed=123)
f <- function(x) x^2
batchMap(reg, f, 1:10)
submitJobs(reg)

## 2014-05-15 23:44:42: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-05-15 23:44:42
## Setting seed: 123
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/01/1-result.RData
## 2014-05-15 23:44:42: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:42: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-05-15 23:44:42
## Setting seed: 124
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/02/2-result.RData
## 2014-05-15 23:44:42: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:42: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-05-15 23:44:42
## Setting seed: 125
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/03/3-result.RData
## 2014-05-15 23:44:42: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:42: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-05-15 23:44:42
## Setting seed: 126
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/04/4-result.RData
## 2014-05-15 23:44:42: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:43: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-05-15 23:44:43
## Setting seed: 127
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/05/5-result.RData
## 2014-05-15 23:44:43: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:43: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-05-15 23:44:43
## Setting seed: 128
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/06/6-result.RData
## 2014-05-15 23:44:43: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:43: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-05-15 23:44:43
## Setting seed: 129
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/07/7-result.RData
## 2014-05-15 23:44:43: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:43: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-05-15 23:44:43
## Setting seed: 130
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/08/8-result.RData
## 2014-05-15 23:44:43: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:43: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-05-15 23:44:43
## Setting seed: 131
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/09/9-result.RData
## 2014-05-15 23:44:43: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:
## 2014-05-15 23:44:43: Starting job on node michels-macbook-air.fritz.box.
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /private/tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-05-15 23:44:43
## Setting seed: 132
## Writing result file: /var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//RtmpWsyc0x/file2dfc355abe4e/jobs/10/10-result.RData
## 2014-05-15 23:44:43: All done.
## Setting work back to: /private/tmp/pkgdocs
## Memory usage according to gc:

# should show 10 submitted jobs, which are all done.
showStatus(reg)

## Status for 10 jobs at 2014-05-15 23:44:43
## Submitted: 10 (100.00%)
## Started:   10 (100.00%)
## Running:    0 (  0.00%)
## Done:      10 (100.00%)
## Errors:     0 (  0.00%)
## Expired:    0 (  0.00%)
## Time: min=0.00s avg=0.00s max=0.00s

[Package BatchJobs version 1.3 Index]